jidea-157 Support ISO codes as country parameter to daedalus (#38)
If the country parameter passed to the daedalus method is a string, a
new helper `country_name_from_arg` checks if it matches either iso2 or
iso3 style codes (2 or 3 uppercase letters). If so, it looks up and
returns the country name which the rest of the daedalus method works
with. An error is thrown if the code is not supported.